% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_fetch_pubmed_records.R
\name{batch_fetch_pubmed_records}
\alias{batch_fetch_pubmed_records}
\title{Get PMIDs and full PubMed records}
\usage{
batch_fetch_pubmed_records(query_string, batch_size = 100)
}
\arguments{
\item{query_string}{Character vector of one or more PubMed query strings. Generally the output of \code{\link{create_query_strings}}.}

\item{batch_size}{Number of DOIs to include per batch. Default = 100.}
}
\value{
Character vector of length number of batches (length \code{query_string}) of unparsed XML with root node "PubmedArticleSet"
}
\description{
Get PMIDs and full PubMed records.
NOTE: silently drop DOIs without PMIDs (i.e., DOIs *not* in PubMed)
NOTE: Output is unparsed XML saved as ".txt" files, so that user can choose to parse with `XML` or `xml2` packages.
}
\examples{

# Fetch a single PubMed batch
my_query_single <- '"10.3389/fpsyt.2018.00207"[DOI] OR "10.1186/s40779-018-0166-5"[DOI] OR "10.1186/s12959-018-0173-5"[DOI] OR "10.1103/PhysRevD.97.096016"[DOI] OR "10.1038/d41586-018-05113-0"[DOI]'

\dontrun{
batch_fetch_pubmed_records(my_query_single)
}

# Fetch a multiple PubMed batches
my_query_multi <- c('"\"10.3389/fpsyt.2018.00207\"[DOI] OR \"10.1186/s40779-018-0166-5\"[DOI] OR \"10.1186/s12959-018-0173-5\"[DOI]"', '"\"10.1103/PhysRevD.97.096016\"[DOI] OR \"10.1038/d41586-018-05113-0\"[DOI]"')
\dontrun{
batch_fetch_pubmed_records(my_query_multi, batch_size = 3)
}
}
