% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_update_esearch.R
\name{check_update_esearch}
\alias{check_update_esearch}
\title{Check whether esearch still on server and if not re-search}
\usage{
check_update_esearch(esearch)
}
\arguments{
\item{esearch}{Esearch object. Generally the output of \code{\link{search_get_pmids}}.}
}
\value{
Esearch object
}
\description{
History kept on eutilities servers for less than an hour, so for large queries needs to be updated.
}
\examples{
\dontrun{
records_search <- search_get_pmids(term = "aquilegia[TITLE]",
                                   dir = NULL, output = NULL)
records_search <- check_update_esearch(records_search)
}
}
