% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pmids_to_dois.R
\name{convert_pmids_to_dois}
\alias{convert_pmids_to_dois}
\title{Convert PMIDs to DOIs}
\usage{
convert_pmids_to_dois(pmids, batch_size = 100)
}
\arguments{
\item{pmids}{Character vector of PMIDs.}

\item{batch_size}{Max number of DOIs to be fetched. Default = 100.}
}
\value{
Tibble of character DOIs and numeric PMIDs
}
\description{
The [NLM ID Converter](https://www.ncbi.nlm.nih.gov/pmc/pmctopmid/#converter) is restricted to records in the PMC corpus, as are tools built on its [API](https://www.ncbi.nlm.nih.gov/pmc/tools/id-converter-api/) such as [rcrossref::id_converter](https://docs.ropensci.org/rcrossref/reference/id_converter.html).
}
\examples{
my_pmids <- c("29904359", "29855634", "29848381", "29805157")

# Returns all available PMIDs (3 PMIDs)
convert_pmids_to_dois(my_pmids)

# Returns number of PMIDs within \code{batch_size} (2 PMIDs)
convert_pmids_to_dois(my_pmids, batch_size = 3)

}
