% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_strings.R
\name{create_query_strings}
\alias{create_query_strings}
\title{Create PubMed query strings}
\usage{
create_query_strings(dois, terms = NULL)
}
\arguments{
\item{dois}{List or character vector of DOIs. For multiple DOI batches, use a list of character vectors. For a single doi batch, use a character vector or list (of length 1) of character vectors. Generally the output of \code{\link{batch_dois}}}

\item{term}{OPTIONAL. Character string of additional terms to append to each query string. Include appropriate booleans. See \href{https://www.ncbi.nlm.nih.gov/books/NBK3827/#_pubmedhelp_Search_Field_Descriptions_and_}{PubMed Help} for available tags.}
}
\value{
Character vector of length number of batches (length \code{dois} if \code{dois} is a list, else 1)
}
\description{
Create PubMed query strings for each DOI batch
}
\details{
NOTE: I understand "query string" and "search string" to be synonymous. I use "query string" for internal consistency.
}
\examples{

# With a single batch of DOIs as a character vector
my_dois <- c("10.3389/fpsyt.2018.00207", "10.1186/s40779-018-0166-5", "10.1186/s12959-018-0173-5", "10.1103/PhysRevD.97.096016", "10.1038/d41586-018-05113-0")

create_query_strings(my_dois)
create_query_strings(my_dois, term  = 'AND "clinical trial"[Publication Type]')

# With multiple batches of DOIs as a list of character vectors
my_dois_list <- list(c("10.3389/fpsyt.2018.00207", "10.1186/s40779-018-0166-5", "10.1186/s12959-018-0173-5"), c("10.1103/PhysRevD.97.096016", "10.1038/d41586-018-05113-0"))

create_query_strings(my_dois_list)

}
