% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatcat_convert.R
\name{fatcat_convert}
\alias{fatcat_convert}
\title{Convert between PMIDs and DOIs using the [Fatcat](https://fatcat.wiki/) [API](https://api.fatcat.wiki/)}
\usage{
fatcat_convert(id, type, quiet = TRUE)
}
\arguments{
\item{type}{Character. Either "doi" or "pmid"}

\item{quiet}{logical. Should success/error info be displayed? Default is TRUE.}

\item{is}{PMID (character) or DOI (character or numeric)}
}
\value{
One row dataframe with DOI and PMID. If `id` not found in Fatcat, returns `id` and NA for other id type.
}
\description{
Non-vectorized function. Wrap in `purrr::map_dfr()` to call on multiple ids. May further develop vectorized option with `purrr::map()` or `plyr::llply`.
}
\examples{
fatcat_convert(32786189, "pmid")
purrr::map_dfr(
  c("10.1056/NEJMoa1916623", "10.1038/ncomms11393", "10.5348/D01-2017-21-OA-1"),
  fatcat_convert,
  type = "doi"
)
}
