% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_batch.R
\name{fetch_batch}
\alias{fetch_batch}
\title{Fetch a batch of records}
\usage{
fetch_batch(
  esearch,
  batch_start = 1,
  records_max = NULL,
  batch_size = 10000,
  save_file = TRUE,
  dir = here::here(),
  file_name = esearch$QueryTranslation,
  quiet = FALSE
)
}
\arguments{
\item{esearch}{Esearch object. Generally the output of \code{\link{search_get_pmids}} or \code{\link{check_update_esearch}}.}

\item{batch_start}{Record batch should start with. Default to 1.}

\item{records_max}{Max records to fetch. Defaults to NULL and dynamically set to max of esearch query.}

\item{batch_size}{Maximum batch size. Defaults to 10000, which is eutilities' limit.}

\item{save_file}{Whether batch should be saved as raw characters to file. Defaults to TRUE.}

\item{dir}{Directory for saving files. Default to project root (\code{here::here()})}

\item{file_name}{Root for file names. Default to esearch query.}

\item{quiet}{Whether to silence messages in console. Defaults to FALSE.}
}
\value{
Efetch object as raw characters
}
\description{
Batch IDs to account for Eutils server limits
}
\examples{
# Run search and don't save any output
records_search <- search_get_pmids(term = "aquilegia[TITLE]",
                                   dir = NULL, output = NULL)
batch <- fetch_batch(records_search)

}
