% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_authors.R
\docType{data}
\name{parse_authors}
\alias{parse_authors}
\title{Extract info about authors from PubMed article}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
parse_authors
}
\arguments{
\item{pubmed_article}{pubmed_article Unparsed or parsed XML of class PubmedArticle. Generally called by \code{\link{parse_pubmed_article_set}} to parse articles within article set.}
}
\value{
Dataframe of information about authors.
}
\description{
Extract info about authors from a single PubMed article. See [MEDLINE PubMed XML Elements](https://www.nlm.nih.gov/bsd/licensee/elements_descriptions.html) for details on elements.
}
\keyword{datasets}
