% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_databanks.R
\name{parse_databanks}
\alias{parse_databanks}
\title{Extract info about databanks from PubMed article}
\usage{
parse_databanks(pubmed_article)
}
\arguments{
\item{pubmed_article}{pubmed_article Unparsed or parsed XML of class PubmedArticle. Generally called by \code{\link{parse_pubmed_article_set}} to parse articles within article set.}
}
\value{
Dataframe of information about databanks. One row per databank.
}
\description{
Extract info about databanks from a single PubMed article. See [MEDLINE PubMed XML Elements](https://www.nlm.nih.gov/bsd/licensee/elements_descriptions.html) for details on elements. See [MEDLINE Databank Sources] (https://www.nlm.nih.gov/bsd/medline_databank_source.html) for details on databanks.
}
