% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_doc.R
\name{parse_doc}
\alias{parse_doc}
\title{Read and parse doc of PubMed records and extract specified datatypes as csv's}
\usage{
parse_doc(
  doc,
  pmids = NULL,
  datatypes = c("table", "abstract", "databanks", "authors", "mesh", "keywords",
    "pubtypes"),
  file_name = "pubmed",
  suffix = NULL,
  dir = here::here(),
  subdir = dir,
  quiet = FALSE,
  return = TRUE
)
}
\arguments{
\item{doc}{Filepath of doc of batch of unparsed PubMed records, such as the output of \code{\link{fetch_batch}}.}

\item{pmids}{Vector of pmids. If pmids not user-provided, pmids will be saved as .rds.}

\item{datatypes}{Types of data to extract from xml for which there is a corresponding "pubmed_" function ("table", "abstract", "databanks", "authors", "mesh", "keywords", "pubtypes")}

\item{file_name}{and @param suffix If both are equal to default, \code{doc} is checked against filename patterns of files generated by \code{\link{fetch_batch}} (either "YYYY-MM-DD_{FILENAME}_{SUFFIX}.txt" or "YYYY-MM-DD_{FILENAME}.txt"). Available values are extracted.}

\item{suffix}{Suffix for file names. For example, record numbers. Defaults to NULL.}

\item{dir}{Directory for saving files (log file and pmids.rds, and extracted csv's, depending on \code{subdir}). Defaults to project root (\code{here::here()})}

\item{subdir}{Directory for saving extracted csv's. Defaults to \code{dir}.}

\item{quiet}{Whether to silence messages in console. Defaults to FALSE.}

\item{return}{Whether to return parsed xml. Defaults to TRUE. Set to FALSE if interested in only side-effect csv's.}
}
\value{
Parsed xml with names = pmids. Also, side-effect of specified datatypes as csv's.
}
\description{
Read and parse doc of PubMed records and extract specified datatypes as csv's
}
