% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_publication_types.R
\docType{data}
\name{parse_publication_types}
\alias{parse_publication_types}
\title{Extract info about publication types from PubMed article}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
parse_publication_types
}
\arguments{
\item{pubmed_article}{pubmed_article Unparsed or parsed XML of class PubmedArticle. Generally called by \code{\link{parse_pubmed_article_set}} to parse articles within article set.}
}
\value{
Dataframe of information about publication types
}
\description{
Extract info about publication types from a single PubMed article. See [MEDLINE PubMed XML Elements](https://www.nlm.nih.gov/bsd/licensee/elements_descriptions.html) for details on elements. See [PubMed Help: Publication Types](https://www.ncbi.nlm.nih.gov/books/NBK3827/table/pubmedhelp.T.publication_types/) and [2020 MeSH Pubtypes](https://www.nlm.nih.gov/mesh/pubtypes.html) for details.
}
\keyword{datasets}
