% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pubmed_article.R
\name{parse_pubmed_article}
\alias{parse_pubmed_article}
\title{Extract summary info about PubMed article}
\usage{
parse_pubmed_article(pubmed_article)
}
\arguments{
\item{pubmed_article}{Unparsed or parsed XML of class PubmedArticle. Generally called by \code{\link{parse_pubmed_article_set}} to parse articles within article set.}
}
\value{
Dataframe of one row with summary information about article.
}
\description{
Extract summary info about a single PubMed article. Extracts only elements which are singular and not more complex elements such as authors. See [MEDLINE PubMed XML Elements](https://www.nlm.nih.gov/bsd/licensee/elements_descriptions.html) for details on elements.
}
