% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pubmed_article_set.R
\name{parse_pubmed_article_set}
\alias{parse_pubmed_article_set}
\title{Extract summary info about PubMed article set}
\usage{
parse_pubmed_article_set(pubmed_article_set)
}
\arguments{
\item{pubmed_article_set}{Unparsed or parsed XML of class PubmedArticleSet comprising one or more PubmedArticle. Generally the output of \code{\link{batch_fetch_pubmed_records}}}
}
\value{
Dataframe of summary information about each article in set. One row per article.
}
\description{
Extract summary info about PubMed article set, containing one or more articles. Extracts only elements which are singular and not more complex elements such as authors. See [MEDLINE PubMed XML Elements](https://www.nlm.nih.gov/bsd/licensee/elements_descriptions.html) for details on elements.
}
