% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_get_pmids.R
\name{search_get_pmids}
\alias{search_get_pmids}
\title{Run E-utilities search and save PMIDs}
\usage{
search_get_pmids(
  term,
  file_name = NULL,
  dir = here::here(),
  output = c("rda", "rds", "txt")
)
}
\arguments{
\item{term}{Query for search}

\item{file_name}{Root for file names. Default to NULL and dynamically set to esearch query.}

\item{dir}{Directory for saving files. Default to project root (\code{here::here()})}

\item{output}{Save full esearch record (as "rds") and/or PMIDs (as "rda" and/or "txt". Default to all (\code{c("rda", "rds", "txt")})}
}
\value{
Esearch object
}
\description{
Batch IDs to account for Eutils server limits
}
\examples{
# Run search and don't save any output
records_search <- search_get_pmids(term = "aquilegia[TITLE]",
                                   dir = NULL, output = NULL)

}
