% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pubmed_files.R
\name{write_pubmed_files}
\alias{write_pubmed_files}
\title{Save PubMed files}
\usage{
write_pubmed_files(pubmed_data, dir = getwd(), prefix = "")
}
\arguments{
\item{pubmed_data}{Character vector of unparsed XML with root node "PubmedArticleSet". Generally the output of \code{\link{batch_fetch_pubmed_records}}.}

\item{dir}{Directory to save files to. Default = current directory via \code{getwd()}.}
}
\value{
SIDE-EFFECT. Unparsed XML in ".txt" files with batch number
}
\description{
Save PubMed files as unparsed XML in ".txt" files, so that user can choose to parse with `XML` or `xml2` packages.
}
\examples{

my_dois <- c("10.1186/s12970-020-0336-1", "10.1016/S0140-6736(19)32971-X", "10.1056/NEJMoa1905239", "10.1371/journal.pone.0226893", "10.1016/S2352-3026(19)30207-8")

batch_dois() \%>\%
create_query_strings() \%>\%
batch_fetch_pubmed_records() \%>\%
write_pubmed_files()
}
