% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_batch.R
\name{parse_batch}
\alias{parse_batch}
\title{Parse batch of PubMed records and extract specified datatypes as csv's}
\usage{
parse_batch(
  batch,
  pmids = NULL,
  datatypes = c("table", "abstract", "databanks", "authors", "mesh", "keywords",
    "pubtypes"),
  file_name = "pubmed",
  suffix = NULL,
  dir = here::here(),
  quiet = FALSE,
  return = TRUE
)
}
\arguments{
\item{batch}{Batch of unparsed PubMed records, such as the output of \code{\link{fetch_batch}}.}

\item{pmids}{Vector of pmids. If pmids not user-provided, pmids will be saved as .rds.}

\item{datatypes}{Types of data to extract from xml for which there is a corresponding "pubmed_" function ("table", "abstract", "databanks", "authors", "mesh", "keywords", "pubtypes")}

\item{file_name}{Root for file names. Defaults to "pubmed".}

\item{suffix}{Suffice for file names. For example, record numbers. Default to null.}

\item{dir}{Directory for saving files. Defaults to project root (\code{here::here()})}

\item{quiet}{Whether to silence messages in console. Defaults to FALSE.}

\item{return}{Whether to return parsed xml. Defaults to TRUE. Set to FALSE if interested in only side-effect csv's.}
}
\value{
Parsed xml with names = pmids. Also, side-effect of specified datatypes as csv's.
}
\description{
Parse batch of PubMed records and extract specified datatypes as csv's
}
